// src/types/mathquill.d.ts
    export namespace v3 {
        type HandlersWithDirection = v1.HandlersWithDirection;
        type HandlersWithoutDirection = v1.HandlersWithoutDirection;
        type HandlerOptions = v1.HandlerOptions<BaseMathQuill>;
        type EmbedOptions = v1.EmbedOptions;
        type EmbedOptionsData = v1.EmbedOptionsData;

        type Config = Omit<v1.Config, 'substituteKeyboardEvents' | 'handlers'> & {
            handlers?: HandlerOptions;
        };

        interface BaseMathQuill {
            id: number;
            data: { [key: string]: any };
            revert: () => HTMLElement;
            latex(latex: string): BaseMathQuill;
            latex(): string;
            reflow: () => void;
            el: () => HTMLElement;
            getAriaLabel(): string;
            setAriaLabel(str: string): BaseMathQuill;
            html: () => string;
            mathspeak: () => string;
            text(): string;
        }

        export interface EditableMathQuill {
            id: number;
            data: { [key: string]: any };
            revert: () => HTMLElement;
            latex(latex: string): EditableMathQuill;
            latex(): string;
            reflow: () => void;
            el: () => HTMLElement;
            getAriaLabel(): string;
            setAriaLabel(str: string): EditableMathQuill;
            html: () => string;
            mathspeak: () => string;
            text(): string;

            select: () => EditableMathQuill;
            moveToRightEnd: () => EditableMathQuill;
            moveToLeftEnd: () => EditableMathQuill;
            cmd: (latex: string) => EditableMathQuill;
            write: (latex: string) => EditableMathQuill;
            keystroke: (key: string, evt?: KeyboardEvent) => EditableMathQuill;
            typedText: (text: string) => EditableMathQuill;
            clearSelection: () => EditableMathQuill;
            blur: () => EditableMathQuill;
            focus: () => EditableMathQuill;
            getAriaPostLabel: () => string;
            setAriaPostLabel: (str: string, timeout?: number) => EditableMathQuill;
            ignoreNextMousedown: (func: () => boolean) => EditableMathQuill;
            clickAt: (x: number, y: number, el: HTMLElement) => EditableMathQuill;

            __controller: Controller;
            R: 1;
            L: -1;
        }

        interface API {
            (el: HTMLElement): BaseMathQuill | null;

            StaticMath(el: null | undefined): null;
            StaticMath(el: HTMLElement, config?: Config): BaseMathQuill;

            MathField(el: null | undefined): null;
            MathField(el: HTMLElement, config?: Config): EditableMathQuill;

            InnerMathField(el: null | undefined): null;
            InnerMathField(el: HTMLElement, config?: Config): EditableMathQuill;

            TextField?: {
                (el: null | undefined): null;
                (el: HTMLElement, config?: Config): EditableMathQuill;
            };

            L: -1;
            R: 1;
            config(options: Config): void;
            registerEmbed(
                name: string,
                options: (data: v1.EmbedOptionsData) => v1.EmbedOptions
            ): void;
        }
    }
// ...
    interface DefaultJquery {
        (el: HTMLElement): DefaultJquery;
        length: number;
        [index: number]: HTMLElement | undefined;
        hasClass: (text: string) => boolean;
    }
    export interface Controller {
        cursor: Cursor;
        textarea: HTMLTextAreaElement;
        textareaSpan: HTMLSpanElement;
    }
    export interface Cursor {
        "-1": MQNode;
        "1": MQNode;
        cursorElement: HTMLElement;
        deleteSelection: () => void;
        insRightOf: (node: MQNode) => void;
        insLeftOf: (node: MQNode) => void;
        show: () => void;
        selection: string[];
    }
    export interface MQNode {
        jQ: DefaultJquery;
        "-1": MQNode;
        "1": MQNode;
    }